+++ src/screens/ContentScreen.tsx
@@ -1,11 +1,11 @@
-import { SafeAreaView, ScrollView, StyleSheet } from 'react-native';
+import { SafeAreaView, ScrollView, StyleSheet, View } from 'react-native';
 import { StatusBar } from 'expo-status-bar';
 import ContentMain from '@components/organisms/Section/ContentMain';
 import ContentAmi from '@components/organisms/Section/ContentAmi';
 import ContentProgram from '@components/organisms/Section/ContentProgram';
 import ContentBannerCarousel from '@components/molecules/Banner/ContentBannerCarousel';
 import ContentPayment from '@components/molecules/Section/ContentPayment';
-import StyledButton from '@/components/atoms/Button/StyledButton';
+import BasicButton from '@/components/atoms/Button/BasicButton';
 
 export default function ContentScreen() {
   return (
@@ -17,9 +17,11 @@ export default function ContentScreen() {
         <ContentAmi />
         <ContentProgram />
         <ContentPayment />
-        {/* 테스트용 코드 */}
-        <StyledButton type='full' content='Apply' active onPress={() => {}} />
       </ScrollView>
+      {/* 테스트용 코드 */}
+      <View style={{ height: 50 }}>
+        <BasicButton content='Apply' round onPress={() => {}} />
+      </View>
     </SafeAreaView>
   );
 }
